Add mixed-precision ONNX export and dtype fixes
Ensure dtype consistency for mixed-precision/ONNX workflows and refactor ONNX export pipeline. Key changes:

- codon/block/embedding.py: Cast cached cos/sin tensors to the input dtype in RotaryEmbedding and InterleavedRotaryEmbedding to preserve mixed-precision graphs (e.g. FP16 exports).
- codon/ops/attention.py: Match attention_mask to query dtype (not forced to float32) and create typed -inf/0 tensors to keep device/dtype consistency.
- codon/utils/onnx.py: Compute RMSNorm variance in FP32 then cast back to input dtype to improve numerical stability for FP16 graphs.
- codon/motif/onnx.py: Major refactor of ONNX export: separate builders (_build_export_args, _export_raw), add _simplify_model and _save_model helpers, and new export() supporting precision ('fp32'/'fp16'), optional simplification (onnxsim), and externalized weights. Dummy inputs now use the model float dtype.
- requirement.txt renamed to requirements.txt and added onnx-related tooling (onnxsim, onnxscript, onnx-simplifier, onnxconverter_common); setup.py updated to read requirements.txt.
- test_prompt.py: Auto-detect graph float dtype from ORT metadata and use matching numpy dtypes for mask and past_kv arrays; added helper _np_float_dtype.
- test_remote.py: Update model_id from 'MotifA1' to 'Motif-A1'.

These changes improve type-consistency, stability, and provide a more flexible, robust ONNX export flow (including FP16 support and optional simplification).

+++ codon/block/embedding.py
@@ -225,6 +225,12 @@ class RotaryEmbedding(BasicRotaryEmbedding):
             cos = cos.repeat(*([1] * (cos.ndim - 1)), multiplier)
             sin = sin.repeat(*([1] * (sin.ndim - 1)), multiplier)
 
+        # Cast cos/sin to match input dtype to keep mixed-precision graphs (e.g. ONNX FP16
+        # exports) type-consistent. cos_cached/sin_cached are non-persistent buffers and
+        # therefore not affected by `model.half()`.
+        cos = cos.to(x.dtype)
+        sin = sin.to(x.dtype)
+
         return (x * cos) + (self._rotate_half(x) * sin)
 
 
@@ -352,5 +358,11 @@ class InterleavedRotaryEmbedding(BasicRotaryEmbedding):
             multiplier = x.shape[-1] // cos_all.shape[-1]
             cos_all = cos_all.repeat(*([1] * (cos_all.ndim - 1)), multiplier)
             sin_all = sin_all.repeat(*([1] * (sin_all.ndim - 1)), multiplier)
-        
+
+        # Cast cos/sin to match input dtype to keep mixed-precision graphs (e.g. ONNX FP16
+        # exports) type-consistent. cos_cached/sin_cached are non-persistent buffers and
+        # therefore not affected by `model.half()`.
+        cos_all = cos_all.to(x.dtype)
+        sin_all = sin_all.to(x.dtype)
+
         return (x * cos_all) + (self._rotate_half(x) * sin_all)
